requirements.txt
django>=5.1,<5.2
django-ninja>=1.3,<2.0
python-telegram-bot==22.7
python-dotenv==1.0.1
openai==2.28.0
